♻️ [Refactor]: create study entity 생성
