+++ src/entities/study/create/api/create-study.api.ts
@@ -1,10 +1,10 @@
+import { fetchAPI } from "@/shared/api/lib";
 import { API_ENDPOINT } from "@/shared/constants/api-end-point";
 import {
   CreateStudyRequest,
   CreateStudyResponse,
   UploadImageRequest,
-} from "../types/create-study.api";
-import { fetchAPI } from "@/shared/api/lib";
+} from "./create-study.api.type";
 
 function sanitizeCodeLikeLinesWithEscape(md: string): string {
   const CODE_LIKE_KEYWORDS = [
