+++ src/entities/study/create/api/create-study.api.type.ts
