+++ src/entities/study/create/api/create-study.type.ts
