+++ src/features/create-study/form/CreateStudyForm.tsx
@@ -8,8 +8,8 @@ import CreateStudyForm2 from "@/features/create-study/form/CreateStudyForm2";
 import CreateStudyForm3 from "@/features/create-study/form/CreateStudyForm3";
 import CreateStudyForm4 from "@/features/create-study/form/CreateStudyForm4";
 
+import { StudyFormInterface } from "@/entities/study/create/api/create-study.type";
 import useStudyStepProtection from "../hooks/useStudyStepProtection";
-import { StudyFormInterface } from "../types/create-study.type";
 
 const CreateStudyForm = () => {
   const searchParams = useSearchParams();
