+++ src/features/create-study/ui/ContentAdderBox.tsx
