+++ src/features/create-study/ui/ContentInputBox.tsx
