+++ src/features/create-study/ui/StudyBasicInfo.tsx
