+++ src/features/create-study/ui/StudyBenefits.tsx
