+++ src/features/create-study/ui/StudyCurriculum.tsx
