+++ src/features/create-study/ui/StudyDescription.tsx
