+++ src/features/create-study/ui/StudyMethod.tsx
