+++ src/features/create-study/ui/StudyRules.tsx
